pdfwrite - gracefully handle inappropriate PDFA command line
If a user sets -dPDFA, but does not set ColorConversionStrategy, and
then processes an input file which has a DeviceN space with an alternate
space which does not match the ProcessColorModel, then we would write
an invalid PDF file (invalid function for the colour space).

Now technically this shouldn't happen, but we can guard against this
specific user error by spotting the problem and returning a rangecheck
error instead, which will cause the earlier code to fall back to
writing a device colour instead.

No diffrences expected.