txtwrite - slacken a restriction on minimum width selection
Bug #699078 " Extracted text is not properly seperated"

Allow the minimum size assigned to the monospaced output to be smaller
than previously (previously we limited it to .75 of the smallest point
size of any used font).

This will reduce the number of horizontal text collisions at the cost
of potentially making the text output wider.

Essentially this is caused by using a monospaced font to represent the
original proportionally spaced input. A sufficiently long run of
characters which are significantly less than the M width could result in
the position being further to the right (when converted to monspaced
text) than the neext horizontal fragment. So we wouldn't try to
space it any further.

Previously we limited the calculation of the minimum character size in
a run to 3/4 of the M space. Here we reduce that to 1/2 which will
permit more flexibility, but as stated, at the potential cost of the
output text being 'wider' (containing more spaces).