txtwrite - better dynamic resizing
A better fix for bug #699078, this implemnets the thought I had while
writing up the description of the problem. In this case we check whether
the current 'effective size' of the monospaced font would lead to
the end of the current run of text lying to the right of the 'next' run
of text on this line. If it would, then use the newly calculated
effective size, no matter how small it is, to avoid collisions.