Fix CCITT Fax Encoder
Bug #696413 "Seg faults found by fuzzing in pdf_store_page_resources (gdevpdfu.c:1663)"

As noted it eh bug report, the problem is that the encoding buffer is
allocated with a size based on the maximum number of bits for horizontal
encoding (6 bits per pixel), but for 2D encoding some of the bits might
use vertical encoding. Vertically encoded bits with an offset of 3 use
7 bits per pixel, not 6.

If we get a line with sufficient 7-bit encodings, where the horizontal
bits are encoded with 6 bits (eg runs of 1 pixel), then we can end up
running off the end of the buffer. Make the allocation assume a maximum
of 7 encoded bits per pixel instead.