CCITT Fax - ensure calculation rounds up
Still Bug #696413, follows on from commit 0221c9

When calculating the number of bytes required to store the encoded lines
we multiply the number of columns byt the number of bits required
for 2 pixels, then divide by 16 to get the required number of bytes.

This commit simply ensures that the division by 16 is rounded *up* and
not down. Previously we could end up dropping up to 7 bits of required
space. Generally this is not likely to be a problem but it is technically
a problem, and can be demonstrated on pathological cases.