Avoid total image dropout in interplated rescales.
Tests show that the following command causes images to disappear:

 gxps -o out.ppm -sDEVICE=ppmraw -72 tests_private/xps/xpsfts-a4/fts_27xx.xps

This is because we scale a 1600 wide image down to 27 pixels across.
As such the weights for each individual pixel round to zero, and we
get nothing displayed.

Adjust for this by still calculating weights as doubles, and carrying
the error across the line. This keeps the total sum of the differences
correct.