pdfwrite - avoid use of /.notdef when converting PCL fonts to type 3
Bug #699102 "using -sDEVICE=pdfwrite option generate an incorrect code 128 barcode"

This is actually an Acrobat bug, every other viewer I've tested works as
expected, Acrobat does not.

The problem appears to be related to defining a /.notdef glyph and
encoding it at a position other than 0 in the Encoding. If we do that,
and use the encoded /.notdef glyph, Acrobat quietly replaces it with the
glyph encoded at position 0 in the Encoding of the font.

Because of the way that PCL fonts work, when we are converting PCL
fonts to type 3 PDF fonts, the character (PCL) 0x1D gets assigned the
glyph name /.notdef. The actual Encoding of the glyph is at position
0x1D however, which causes Acrobat to substitute the glyph at Encoding
position 0.

This is obviously wrong. However, since its Acrobat, here we add yet
another heuristic to avoid defining glyphs with the name /.notdef when
creating type 3 PDF fonts from PCL fonts.

No differences expected