pdfwrite - reject inappropriate ICC profiles
PDF files cannot include the full possible range of ICC profiles, if
we encounter one of these (eg from an XPS file) then refuse to embed it
and instead return gs_error_rangecheck, so that the calling code will
fall back to a device space representation.

NOTE we currently permit GS_UNDEFINED as a valid data_cs which we really
shouldn't but the current code, when converting a CIEBasedDEF or
CIEBasedDEFG (possibly other CIE spaces) to ICC creates a profile
where data_cs is GS_UNDEFINED (which is 0, so probably not actually
filled in).