pdfwrite - don't write device spaces for masks representing glyphs
This came up while looking into ICC profile embedding. If we are
processing a mask bitmap which represents a glyph (eg when creating a
type 3 font for a font which cannot be embedded directly in a PDF) we
do not want to write a device colour.

We probably *never* want to write a device colour in here, but I'm not
certain of that, I am certain that the high level colour will already
have been set if we are processing text though. So in this commit, limit
the change so that we don't write device colours for glyphs.