Simplify image enumerator rectangle logic for scalers.
We store the image size as a rectangle in the image enumerator.
For the purposes of scaling, we also store the rectangle that
needs to be rendered. We expand this slightly to allow for
the support requirements of the scaler, thus giving a rectangle
that is really what we need to decode.

In order to correctly setup the scaler we then need to try to
guess how much expansion was done. My previous attempts at
this guesswork failed.

A simpler solution would appear to be to just save both the
render and decode rectangles stored separately. This simplifies
the logic.