pdfwrite - ensure correct ICC profile version being written out
Bug #699104 "Identify ICC profiles unsuitable for embedding in output PDF and fall back to device space"

Previously we did not check the minor version number of ICC profiles
when writing an ICCBased colour space. Since PDF versions below 1.7
support either 4.0.0 0r 4.1.0, this could mean that we wrote a version
of profile that was too great for teh PDF version.

Following on from Michael's commit f732ea3a999f73afa46c2cc4ca28b067f671d29c
we can now determine the minor version number of profiles, and we now
use that to decide whether to use the profile as it is, or drop back to
a version 2 profile.

Since falling back to a version 2 profile can cause subtle changes in
colour, we now make the default CompatibilityLevel for PDF output 1.7.
Some PDFSETTINGS convert all colours to device spaces and these have
not been altered (not required). We could have defaulted to PDF 2.0
output, but the code is rather new, and also prevents certain constructs
(eg TransferFunction in the graphics state).