PDF interpreter - remove /.gs.pdfobj# entries from pdfmarks
Since we increased the places where we use /.gs.pfgobj# to track the
PDF object number, we have been emitting PDF files from pdfwrite
which contain dictionaries with illegal PDF names.

The '#' in PDF naming is used to escape unprintable characters (eg
space), it must be followed by 2 hex digits, ours isn't.

Trying to fix this in the pdfmark processing proved problematic, so
instead, remove it at the PostScritp level.