Fix FAPI's CharString dictionary accessing
Because dictionaries are (generally) made up of sparse arrays (one each for
keys and values), it's not generally safe to simply access indices into those
arrays - which the FAPI code was doing.

Normally, for that type of access, we'd use the dictionary content enumerator
function but, in this case, the FAPI API does not enforce iterating through
the dictionary contents, but allows for "random" accessing the contents.

However, the requirement to handle broken or invalid CharStrings also means we
neatly handle empty slots in the sparse arrays.

So, the only change required here is to use the maximum size of the dictionary
object, rather than the "size" to know how many indices to check.

As an additional safety check, however, also add checks for the key/value
types, and treat an invalid type as an empty slot.