PS interpreter - fix parsing of NChannel Colorants dictionary
The code was previously using dict_index_entry to retrieve entries
from the Colorants dictionary, but it did not account for the function
returning gs_error_undefined when a slot in the dictionary was present
but unused. Since dictionaries are sparse arrays, this is possible.

This commit reworks the code to use dict_first() and dict_next() to
retrieve elements from the dictionary, and starts at the maximum entry
and works down, instead of the previous code which started at 0 and
worked up.