zfapi.c: another case of is_glyph_index set wrongly
This is related to freetype now a) automatically selecting a Unicode cmap for
every font, if one is available, and b) automatically generating a Unicode cmap
for every font when one is not available.

Logic that we had pushed down to the FAPI/FT interface layer, we now need to
apply earlier.