Bug 699160: rejig the .so "loaders" builds.
The two simple .so "loaders" (gsx and gsc) were built as a single step, source
to exe, process. Because of that LDFLAGS wasn't used.

We'll now separate the compile and link stages, meaning we can properly use
CFLAGS and LDFLAGS. It also means we can have more control of the order in
which the -L lib search paths are applied.

Properly pass on debug options, and extend these changes to the PCL/XPS/PDL
.so/exe targets.