Tweak logic for skipping clipped image lines.
As gs runs through image data, it unpacks each line into
a buffer, then checks whether that buffer is trivially
outside the clipping region. If it is, then it doesn't
bother to call the line renderer.

This has 2 implications. Firstly, we unpack the line data
even if it is to be skipped. Secondly, for line renderers
that rely on counting the number of lines they have been
passed (rather than on consulting the position of the dda)
to know where they are, this can cause problems.

Accordingly, this commit rejigs the code a bit. As well
as a function pointer within penum that points to the
rendering routine, we have a new function pointer that
points to an optional "skip_render" routine.

We also rejig the code so that we only unpack the data if
we know it is going to be used.