Another attempt at fixing the 'skip' logic for halftoned images.
When rendering images, if the dda is not shared between the
halftoner and gs, gs cannot just skip lines that it knows should
be clipped, lest the halftoner get confused about it.

The previous attempt to solve this used a separate 'skip_render'
routine, but the decision as to whether to call this or not was
itself dependent on the values of the dda. When scaling down
scanlines can be completely skipped. In the cases where the dda
in gs doesn't match the one in the halftoner exactly we could
get dropouts.

So, a new attempt. Here we introduce a new function call that
checks with the halftoner whether the next scanline will be
required or not. If it is, gs unpacks the data and passes it in.
If not, the data can just be skipped.