jbig2dec: Fix incorrect adaptive pixel location check.
Previously for the adaptive pixel template 2 the check for the
location of the adaptive pixel was wrong. Because generic regions
with nominal locations are more common this lead to the
unoptimized decoder being used. Only a few images where the
adaptive pixel location happened to be at this wrong location
used the optimized decoder, and presumably were decoded
incorrectly. This incorrect location check therefore hid the bug
in the optimized generic region decoder concerning what pixels
are included at what location in the arithmetic integer context.

Now the adaptive pixel location check is corrected as well as the
bug in the optimized generic region decoder itself.