Multiple master font cached glyphs
In the event that the WeightVector is changed (via setweightvector) on an
already defined font object, we had no way to invalidate the glyph cache
entries already created for that font, with the previous WeightVector.

Given that MM fonts generally substitute for an entire use of a "real" font,
and changing "mid stream" is rare (and almost entirely restricted to test
files), this uses the sledgehammer approach of purging the existing glyphs
from the cache.

In real world use, this is going to have less effect on performance than adding
a conditional check on a variable length array of floating point numbers is
likely to have.