Attempt to fix vertical shifts seen in halftoning.
Differences are evident between old (slow, non SSE based) and new
(fast, C/SSE based) halftoning. This is not because of the SSE,
but rather because of different calculations in the C that prepares
the screen for the fast SSE based thresholding, specifically, I
believe, to the calculation of the y offset into the texture tile.

This commit changes the texture tile y offset calculation in the
'fast' code to involve screen_phase.y, in the hopes that this
will better match the 'old' version.

With this fix in, we see just a few PS/PDF diffs, but lots of PCL
diffs (as I guess you'd expect, as PCL makes much more use of this
kind of stuff). I've checked a random selection of them, and in all
cases the new page mode renderings with the "fast" code match the
page mode renderings given by the "old" code.

There are instances where the clist rendering no longer matches the
page mode rendering, but in the tests I've done the clist old code
does match the clist fast code. So we are consistent at least. I
suspect that there is something up with the way the y_phase is
transmitted across the clist, and have opened bug 699209 to
track that.