Add processing for all the PDF interpreter parameters
The PDF interpreter previously used the PostScript environment, we can't
use that now (when not running under PostScript), so we need to process
all the various parameters accepted and used by the interpreter.

This turns out to be a great deal more complicated than expected. For
non-PostScript interpreters the command line is processed in the 'pl'
level, and the parameters are mainly passed to the device, after it is
installed, by the usual param_list. Anything which the device doesn't
process (eg NOPAUSE) is handled 'specially' in the PL code.

Icky, icky, icky. This means that the PL code must process every
parameter for every (non PostScript) interpreter, and store it in its
own structure. Even more icky, there's apparently no defined way to pass
on the PL level parameters to the actual interpreter. I may be missing
something here.

Instead, the PDF interpreter has to use the pl_main_get_instance() call
to pull out the main instance, and then interrogate the parameters
stored there. This is further complicated by the fact that although the
structure is typedef'ed in the .h file, its defined in the plmain.c
file. Which means that other interpreters can't access its members
anyway, even after they get access to it!

So, move the definition of pl_main_instance_s to the plmain.h file,
add the required headers (for arg_list and gs_c_param_list) to the
plmain.h file. This allows the PDF itnerpreter to access the structure.
Obviously remove the definition of the structure from plmain.c.

Add *lots* of new parameters to the main instance to track the many and
various parameters the PDF interpreter needs. Add lots of matching
parameters to the PDF interpreter context. Obviously I don't actually
need to mirror these, I could just pull them direct from the main
instance, but that would mean a function call every time which could
work out expensive.

In pdf_im_process_file, copy the parameters from the main instance to
the interpreter context.

Set the content of the main_stream to NULL before we open it to ensure
that all the members are initialised....