Add a pdf_read_object which reads: x y obj ..... endstream/endobj and turns the sequence into a pdf_obj. When we do that, add it to the cache.
Implement pdf_dict_put.

create a pdf_make_name to make it easier to use pdf_dict_put

Change all refcnt++ and refcnt-- to use pdf_countup and pdf_countdown

Add a function to process the Root dictionary (partially implemented).
If the root dictionary (stored in the trailer dictionary) is a reference,
then replace the entry in the trailer dictionary with the dereferenced
object. Might be nice to write a function for this.