Bug 699214: Add and use a "no cache" flag for glyphs
Previously, to avoid using a cached glyph, the PXL code went around the
graphics library API to set a flag in the show enumerator that really should
only be touched internally.

This adds the capability to the enumerator for the client to request that no
cached glyphs be used for the current block of text, and changes the PXL code
to used it.