Straighten out some reference counting, add some comments to help keep it straight.
The various 'read' routines all return the object on the stack, in all
cases the object has a reference count of 1 reflecting the reference
held by the stack. When the stack is popped the reference is
decremented.

Other routines which return an object directly, ie the 'make' routines
(currently only pdf_make_name) and the 'get' routines (pdf_dict_get
and pdf_array_get) increment the reference count of the object to
reflect the fact that the *caller* now holds a reference to the returned
object.

pdf_dict_put increments the reference count of the key and value objects
*if* they are inserted into the dictionary, otherwise it does not alter
them , reflecting the fact that the only references are held by the
caller (and any other existing references).