Add "transform_pixel_region" device function.
Use this for handling image type 4 rendering.

The default version of this goes to rectangles. It also knows how to
send portrait images of 8bps to copy_color.

Next step is to write versions of this for the memory mapped devices
to know how to go direct.

Essentially we are copying the guts of the
image_render_color_icc_{portrait,landscape,skew} into a default
device method, and making the code that called those functions
call the device method instead.