Alter xref dumping code to be clearer for compressed objects
Initial code to handle reading a stream object. Currently this reads the
Length from the dictionary, seeks to the end of the stream and checks
for an endstream endobj pair. This may be too expensive in performance
terms, so we might want to optionally skip this step in future. Its safe
to do this because we should not get stream objects in comrpessed
streams. We may want to check that too.....

Initial code to use read compressed streams and extract objects from
them (not fully implemented).

Fix a bug in pdf_read_xref_stream_dict which was using stell on a
pdf_stream directly, instead we must use the stream * stored inside.