Fix FlateDecode filter by making min_size 2 O.O
Alter the prototype for pdf_read_object.

Alter pdf_read_object to handle reading objects from compressed object
streams. In this case the object and generation numbers are not followed
by an 'obj' keyword, and there is no trailing endobj.

Add code to check a compressed object stream is an ObjStm and has enough
entries to satisfy the requested object index.

Change the way that 'common' PDF object variables are defined in
the structures. Use a macro for the common variables, this means we
don't need to mess about with "->object." when we are dealing with
specific PDF data types but can address these variables directly for all
types.

Alter the xref entry so that the compressed and uncompressed data are
stored in a union.

Define (not used yet, that's the next job) a structure for the xref
table. Create a PDF object type for it and make it a PDF object so we
can reference count it and apply countup and countdown etc.

Make the PDF stream a PDF object too so we can reference count it as
well (again not yet implemented).

Add lots more debugging when -dPDFDEBUG is set, make the debugging code
produce prettier output.