Add Pages dictionary to context structure (for faster access), count it down when releasing context.
Add code to read and store a reference to, the Pages dictionary.

Fix major screwup in understanding of compressed object streams; all the
object numbers, along with the offset from the first object, are present
at the start of the stream, followed by the actual objects. Put the
pdf_read_object routine back the way it was, we can't use that for
compressed object streams.

Add all the escape, octal and special end of line handling to the string
reading code.