Add a 'num_pages' entry to the context for easy quick reference.
Add a 'pdf_stell' which only operates on the main PDF file, it returns
the current position (from the start of the file) taking into account
any bytes that have been 'unread'.

Fix a reference counting mistake when dereferencing an object from
the cache.

Fix the relinking of the most recently used object; only update its
next and previous members if they are not NULL.

When dereferencing, save and restore the current position of the main
PDF file, so that we can carry on where we left off after the object
has been dereferenced. Only do this for objects not returned from the
cache.

Add a 'pdf_array_put' function to replace elements in arrays.

When creating an indirect reference object, store the object and
generation numbers in the correct members.

When we read the Pages tree, store the number of pages in the context.

New function pdf_page_dict, used to retrieve the page dictionary for a
specific (0-based) page index. This function is recursive in order to
follow the tree structure. When it finds an indirect reference to a
/Pages node, it replaces it with the dereferenced object. We don't do
this for Page (leaf) nodes because they could be large. We may need to
rethink this.

Add a 'pdf_render_page function' to render the page contents of a page
given a (0-based) index.

Have pdf_process_pdf_file render each page in turn.