Bug 699201: Fix alignment of rendering bands memory
The bottom section of the "raster buffer", when rendering through the page
clist, is reserved as a cache (tile cache, to be specific). Hence to bottom
address of the memory we actually render into starts immediately after the
cache.

No account was taken when setting up that cache size of the alignment expected
by the rendering code.

We'll now ensure that the the cache ends, and the raster band memory starts
on an appropriately aligned address.