jbig2dec: Remember to put all new symbols into ref/agg dictionary.
Previously the symbol dictionary propagated to the text region
decoder for a refinement/aggregate coded symbol did not contain
all necessary symbols. Only the symbols in the symbol dictionaries
referred to (SDINSYMS) by the currently parsed symbol dictionary
would be propagated into the text region decoder (SBSYMS).

Table 17 in the specification states that SBSYMS should be set
according to 6.5.8.2.4 which clearly states that SDINSYMS as well
as any symbols decoded up to that point (NSYMSDECODED symbols in
SDNEWSYMS) ought to be introduced into the dictionary that is
propagated to the text region decoder (SBSYMS). This is now done
by providing the text region decoder not just with a single symbol
dictionary, but both with SDINSYMS and SDNEWSYMS allowing the
decoder to refer to symbols from both.