Fix bug 697671: Banding/Page mode mismatch
The simplified case had a character (copy_mono) followed by a rect
fill.

The copy_mono for the character sets the rect.height clamped to the
"cropping_max" (dev->height) for the band in the writer, but it doesn't
actually write the height to the clist for that band. Then the reader
was using the entire bitmap height, but that went past the end of the
device, so the next time the state.rect.height was used in the reader,
it was wrong.