Numerous changes working towards accessing page streams
Add dict_known dict_known_by_key and dict_copy functions.

dict_known sees if a char * describing a key is known in the dictionary.
dict_known_by_key uses a pdf_name * as the key.

dict_copy copies the contents of the source dictionary to the target, if
the target already has any of the key/value pairs they are replaced
with the ones form the source.

Add pdf_merge_dicts. If the source key is *not* known in the target then
copy the source key/value pair to the target. Ignore (do not copy) keys
from the source already present in the target.

if we countdown a NULL pointer, then ignore it (like freeing a NULL
pointer)

Fix a stupid typo in pdf_dict_put. Also, when increasing the size of the
dictionary, actually remember to store the newly created key value
arrays in the dictionary!

Add a pdf_make_dict function to create a new empty dictionary.

Reading beyond the end of file on the stream causes it to close the
underlying file (!!). So rearrange the file size determination and
when reading the header do not attempt to read more bytes than are in
the file.

Modify pdf_get_page_dict so that if any page node defines page
attributes to be inherited, we make a new dictionary and store the
attributes in it. If we inherited any from the node above, do the same
and copy those attributes. When we find the page dictionary, merge it
with the inherited attributes.

Add a place-holder function for checking the existence of transparency
on a page.

Allow for the Info dictionary not to be present (it is optional).