Add the capability to 'mark' and 'cleartomark' the loop detection
This allows us to prevent throwing an error when we (legally) dereference
the same object in two different branches of a Resources dictionary.

Remove the hack for the 'current' object in pdf_dereference_object,
deal with the current object by adding it immediately after creating
the loop detector. This is better because it allows for a lower level
branch referencing the ultimate parent to be detected.

We do not mark/cleartomark around the recursion in pdf_get_page_dict
because the Pages intermediate and leaf nodes can only legally have
a single reference. I think the inheritable attributes are required
to *not* be indirect references (in an intermediate node), but I could
be wrong. Even if they are permitted to be indirect references, there's
no point in having more than a single reference to them in an
intermediate node.

If we ever do decide this is needed then it only needs to be done around
the pdf_get_page_dict recursion.


Add the relevant files to Git....