Add gxdso_copy_color_is_fast.
Default implementation returns 1 if copy_color is not the default
(as the default drops to copy_rects).

Also the lcvd device (as used by pdfwrite and ps2write) returns 0
to stop us calling it in this case. The lcvd device lies about
the width/height, and so clipping goes wrong.