Fix problem with transparency and ROPs.
A comment from 1998 in gdevrop.c indicates that
if we are given a LOP that says "S is transparent", and
S isn't used, then we should ignore that flag. (Similarly
for T). Tests seem to indicate that this is indeed true
(See page 12 of C425.bin for an example).

Unfortunately, when we start 'folding' LOPs down onto simpler
ones, we can start with a LOP where S (or T) matters, and
end up with one where it looks like it doesn't. As such we
introduce another flag so we can avoid trying to remove
the S/T transparency flags after we have started to fold the
rop down.

We use this in mem_gray8_rgb24_strip_copy_rop, and rop_get_run_op
to ensure we don't throw away transparency by mistake.

For example, in mem_gray8_rgb24_strip_copy_rop when called with
lop=0x2fc, with S as a bitmap where both colors are the same, the
current code spots that S is always 1, and rewrites the rop to be
0xff. This in turn leads to T not being required, so it is ignored,
despite the fact that T should still be consulted for transparency.

(Check that page 12 of C425.bin is still OK!)