If we find an error early enough when using an XrefStm from a hybrid file, restart and use the xref table instead. Should probably extend this later. See Bug 695403 for a hybrid file with a broken XrefStm and (at least partially) a valid xref table.
Split xref reading out from file opening, so we can discard an xref
and read a different one.

Add flag to decide whether to prefer XrefStm in a hybrid file, initally
set to true.

Add flag to track whether a file is a hybrid.

When dereferencing, check the xref entry to see if it has an object
number, if it doesn't then its undefined, throw an error.

Fix some compiler warnings