Add array and dictionary marks as PDF objects
Allows us to stop storing pointers to the stack, which doesn't work
if the stack relocates (due to growing in size).

Add functions to put a mark on the stack, count back to a mark, and
clear back to a mark.

Alter the array and dictionary reading code to use the marks instead of
storing a pointer to the stack.

Fix bugs in the calculation of stack size, the detection of a requirement
to grow the stack, the copying of the stack contents on growth.