Changes to Pages tree processing
Up to now we processed the Pages tree lazily, replacing all indirect
references in the tree with dereferenced objects as we used them.
Including any Page dictionaries as we retrieved them.

This is potentially wasteful; page tree dictionaries can be large
(hundreds of entries in the Resources dictionary) and are generally
only used once, to render a page.

With this commit, if we dereference a node and discover its a Page
(rather than Pages) node, ie a leaf, not an intermediate, then we create
a new dictionary with /Type /PageRef and one other key /PageRef which
holds the indirect reference to the actual page dictionary. We then
store that in the Kids array in place of the previous indirect
reference.

However, we pass the Page we dereferenced dictionary to the code which
checks for node type and handles merging the inherited page attributes.
Although at first sight we could simply pass on the PageRef dictionary
and have the node type processing dereference the page dictionary
again, this would be wasteful and, more importantly, would trip the
loop detection, because we'd be processing the same reference twice.

Of course if we do process the pages out of order, then the node
processing code must be prepared to encounter a PageRef node. In this
case we do dereference it to recover the page dictionary which we
return. We will only be dereferencing the object once though, so it
won't be a problem.

This is, of course, a compromise between storage and performance. In the
usual case (processing pages in order) we will only dereference each page
once, and will not store the page dictionaries. In the worst case we
will dereference each page twice. Slightly wasteful of performance
but this is a less usual situation.