Move stack handling to its own file.
Change dictionary and array processing. We now push marks onto the stack
when we see the start of an array pr dictionary, and then call the
relevant routine to manufacture one from the stack element when we
encounter a closing array or dictionary mark.

This means we need to modify pdf_read_object, it can't assume that it
will get a complete object by calling pdf_read_token, but must continue
reading tokens until it finds an 'endobj'. If we find another keyword
before that, then we return an error.

read_xref needs to be able to read the trailer dictionary, so define
a utility routine to do that for it.