Fix quirk in image enumerator creation routines.
When we start an image, the image code runs through the handlers
for different types of image until it finds one that can handle
the required type.

These handlers are queried with the current device in as penum->dev.

The image code then sets up any clip or rop device on top of this
device.

If the handlers were to call device functions, they would be in
the somewhat odd position of calling a different device in the setup
phase than they would be calling in the processing phase. This
upsets a forthcoming commit.

The fix is to ensure that any clip and rop devices are setup BEFORE
the image handlers are called.