Change PDF_TRUE and PDF_FALSE into a PDF_BOOL and pdf_bool data type, so we can do pdf_dict_get_type on boolean objects.
Make dict_from_stack non-static so we can use it from the image code.

Seek to the exact offset of the XrefStm stream before applying a filter
to read it....

Use pdf_tell instead of stell() - 1 to determine the position f the
stream when reading an xref stream dictionary.

Move the check for an already filled in object number when reading xref
tables, so that we actually read the xref entry from the file, even if
we already have an entry for it, but discard it after reading.

When recursively reading xrefs, don't use ctx->is_hybrid to determine
whether the current trailer dictionary has a XrefStm entry, because we
might be processing the Prev from a modified hybrid file, and we've
already read the XRefStm. that is, the file *is* a hybrid, but the
current trailer doesn't have an XRefStm.

When we finish reading an XRefStm, don't assume that we've finished,
check the current trailer dictionary to see if it has a /Prev (this
can only happen in a hybrid file) and process it if we do.

hook up the BI/ID/EI operators to the skeleton support code.