Honor the single object clump flag when allocating strings
When checking whether a given clump was a candidate to allocate a string,
the code was not checking the flag that marks the clump as being for a
single object. That could lead to the clump being freed (with it's original,
single object content, along with the string) when the string(s) were still
in use.

This is better than resetting the flag since single object clumps are such
for a reason (they implement immovable memory), and we *never* want
immovable memory to become movable