Fix decode params dictionary for single filters.
When creating a pdf_stream object, note the original stream which is the
source of data for this stream. When closing a pdf_stream, close all
the streams in the chain until we reach the original filter. Add an
extra (probably redundant) check to ensure we never close the main
stream.

When there are no filters to add, make a new pdf_stream anyway, and
put the original stream * in it. This allows us to close a newly
created stream without worrying about it being a duplicate of the source
stream.

Add support for RunLength, DCT, RL, A85 decompression filters.

Use pdf_close_file to close the stream created for inline images.

Use pdf_close_file to close the streams created for content stream
interpretation, and the streams created for dereferencing compressed
objects.  Close the main stream using sfclose() instead of
pdf_close_file(). Clean up some inconsistency between use of pdf_close_file
and simply freeing the pdf_stream memory.