Spot the existence of an Encrypt entry in the Trailer dictionary and stop processing right there, we can't yet handle encrypted files.
When dereferencing an object in a compressed stream we need to handle
COD objects (arrays and dictionaries) carefully, Objects in compressed
object streams do not have x y obj...endobj sequences so we can't
use the usual object reading, but we do need to process tokens until
we find a matching closing mark. Do this by noting the initial depth
and then process objects until we find a completed COS object, and the
current stack depth is 1 more than we started at (handles nested
objects). Might need to code something better for this eventually.