Tweak previous fix to rounding in image code.
Back in September 2016, commit bffe2011e1 went in, entitled "Fix bug
69707.  SMask subpixel offset must match image offset."

In this, an adjustment was applied to the dda to attempt to make it
match the rounding used in the mono case.

This adjustment was made *after* the position of the first edge was
made, rather than after it, resulting in a slight difference to the
size of the first pixel of a run.

This error was spotted when comparing the output of an experimental
optimisation to use memcpy for 1:1 runs of data.