Increase the data_size for pattern-clist to 128k (from 32K).
The tile cache is kept in the clist buffer, and there are a few files that
trip over the limit of the number of hash slots with the small data_size,
particularly comparefiles/Bug689851.pdf at 72 dpi. With this file, the clist
size is reduced by about 65%. Other than the pattern-clist, the tile cache
seems to work adequately without needing to delete tiles.

Also this means that more data can be accumulated for image_data before
needing to flush the buffer to the memory_clist storage, which should
improve performance a bit (but pattern-clist is never going to be fast).