Copy the initial PDF file filename. We may need to read to EOF and reopen the file if we are repairing it.
Add some missing return checks

Fix a typo in pdf_process_xref_stream

Add code to get the requested Box from the PDF file and set the media
to that box size (apparently does not reset the clip, why ?)

First pass at code to build an xref table by scanning the PDF file
looking for objects.

Only attempt to read the Info and Root (Ctaalog) dictionaries if we
have a Trailer dictionary. If we rebuilt the xref we won't.

Don't attempt to process pages unless we have a Root (Catalog)
dictionary, either read from the Trailer dictionary or picked up when
rebuilding the xref.